requirements.txt
asgiref==3.5.0
Django==4.0.4
django-crispy-forms==1.14.0
django-extra-views==0.14.0
django-funky-sheets==0.2.0
fanstatic==1.2
js.jquery==3.3.1
numpy==1.22.3
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
shutilwhich==1.1.0
six==1.16.0
sqlparse==0.4.2
WebOb==1.8.7
